requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-celery==3.4.1
dbt-core==1.8.0
dbt-postgres==1.8.0